refactor: extract route guard to hook
