+++ pages/_app.tsx
@@ -2,7 +2,7 @@
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useEffect, useMemo } from 'react';
 import { ThemeProvider } from 'next-themes';
 import '@/styles/globals.css';
 import '@/styles/themes/index.css';
@@ -12,16 +12,10 @@ import { ToastProvider } from '@/components/design-system/Toaster';
 import { NotificationProvider } from '@/components/notifications/NotificationProvider';
 import { supabaseBrowser } from '@/lib/supabaseBrowser';
 import { env } from '@/lib/env';
-import { LanguageProvider, useLocale } from '@/lib/locale';
+import { LanguageProvider } from '@/lib/locale';
 import { initIdleTimeout } from '@/utils/idleTimeout';
-import {
-  isGuestOnlyRoute,
-  isPublicRoute,
-  canAccess,
-  requiredRolesFor,
-  getUserRole,
-  type AppRole,
-} from '@/lib/routeAccess';
+import { isGuestOnlyRoute, isPublicRoute } from '@/lib/routeAccess';
+import useRouteGuard from '@/hooks/useRouteGuard';
 
 import { PremiumThemeProvider } from '@/premium-ui/theme/PremiumThemeProvider';
 import { ImpersonationBanner } from '@/components/admin/ImpersonationBanner';
@@ -41,7 +35,6 @@ function GuardSkeleton() {
 }
 
 function InnerApp({ Component, pageProps }: AppProps) {
-  const { setLocale } = useLocale();
   const router = useRouter();
   const pathname = router.pathname;
 
@@ -57,9 +50,6 @@ function InnerApp({ Component, pageProps }: AppProps) {
     [pathname]
   );
   const showLayout = !isPremium && !isAuthPage && !isNoChromeRoute;
-
-  // NEW: treat /pricing as public even if routeAccess hasn't been updated yet
-  const isPricingRoute = useMemo(() => /^\/pricing(\/|$)/.test(pathname), [pathname]);
 
   useEffect(() => {
     const cleanup = initIdleTimeout(env.NEXT_PUBLIC_IDLE_TIMEOUT_MINUTES);
@@ -81,80 +71,7 @@ function InnerApp({ Component, pageProps }: AppProps) {
     return () => sub?.subscription?.unsubscribe();
   }, []);
 
-  const [isChecking, setIsChecking] = useState(true);
-
-  useEffect(() => {
-    if (!router.isReady) return;
-    let active = true;
-
-    (async () => {
-      try {
-        const { data: { session } } = await supabaseBrowser.auth.getSession();
-
-        const guestOnlyR = isGuestOnlyRoute(pathname);
-        const publicR = isPublicRoute(pathname) || isPricingRoute; // NEW: pricing forced public
-
-        // Expired session → sign out, but don't block /pricing
-        const expiresAt = session?.expires_at;
-        if (session && expiresAt && expiresAt <= Date.now() / 1000) {
-          await supabaseBrowser.auth.signOut();
-          if (!guestOnlyR && !publicR) {
-            router.replace('/login');
-            return;
-          }
-        }
-
-        const user = session?.user ?? null;
-        const role: AppRole | null = getUserRole(user);
-        if (!active) return;
-
-        // If unverified, previously forced /auth/verify; allow /pricing to remain open
-        if (user && !user.email_confirmed_at && !user.phone_confirmed_at && !isPricingRoute && pathname !== '/auth/verify') {
-          await supabaseBrowser.auth.signOut();
-          router.replace('/auth/verify');
-          return;
-        }
-
-        if (user) {
-          const { data: profile } = await supabaseBrowser
-            .from('user_profiles')
-            .select('preferred_language')
-            .eq('user_id', user.id)
-            .maybeSingle();
-          const lang = profile?.preferred_language || 'en';
-          setLocale(lang);
-        }
-
-        if (guestOnlyR && user) {
-          const nextParam = router.query.next;
-          const next = typeof nextParam === 'string' ? nextParam : Array.isArray(nextParam) ? nextParam[0] : undefined;
-          router.replace(next || '/dashboard');
-          return;
-        }
-
-        // Role/guard check — skip for /pricing
-        if (!isPricingRoute && !canAccess(pathname, role)) {
-          const need = requiredRolesFor(pathname);
-          if (!role) {
-            router.replace({
-              pathname: '/login',
-              query: { next: pathname, need: Array.isArray(need) ? need.join(',') : need },
-            });
-          } else {
-            router.replace('/403');
-          }
-          return;
-        }
-      } finally {
-        if (active) setIsChecking(false);
-      }
-    })();
-
-    return () => {
-      active = false;
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [pathname, router.isReady, isPricingRoute]);
+  const { isChecking } = useRouteGuard();
 
   // Background token expiry safety — don't force-login on /pricing
   useEffect(() => {
